Revert to CLI version 0.263.0 (#192)
We shipped a regression in `0.264.0` where we started interpolating
resource names before the prefix presets are applied. This PR pins the
default setup-cli back to a good version.

Relevant changelog entry:
* Changed logic for resolving ${resources...} references. Previously
this would be done by terraform at deploy time. Now if it references a
field that is present in the config, it will be done by DABs during
bundle loading (https://github.com/databricks/cli/pull/3370)